+++ src/widgets/ext/menu_list.rs
@@ -36,7 +36,7 @@ pub struct MenuListOptions {
     pub active_shader: Option<ShaderEffect>,
     pub menu_animation: Option<AnimationMachine>,
     pub portal: UiPortalTarget,
-    pub z_index: i16,
+    pub z_index: f32,
     pub action_prefix: Option<String>,
 }
 
@@ -74,7 +74,7 @@ impl Default for MenuListOptions {
             active_shader: None,
             menu_animation: None,
             portal: UiPortalTarget::AppOverlay,
-            z_index: 100,
+            z_index: 100.0,
             action_prefix: None,
         }
     }
